tools/blog-animations/pyproject.toml
[project]
name = "blog-animations"
version = "0.1.0"
description = "Render blog post animations (HTML + SVG) to optimised animated WebP / GIF via Playwright."
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.45",
    "pillow>=10.0",
    "numpy>=2.4.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["blog_animations"]
